multiline input. modularized files
