+++ frontend/components/pokemon/pokepaste.tsx
@@ -2,13 +2,14 @@
 
 import * as React from "react";
 import { usePokePaste } from "@/lib/pokemon/use-poke-paste";
-import { Input, Button, Card, CardHeader, CardBody, CardFooter } from "@nextui-org/react";
+import { Button, Card, CardHeader, CardBody, CardFooter, Textarea } from "@/components/nextui-use-client";
 import Image from "next/image";
 import { StatsTable } from "@pkmn/types";
 
 export default function PokemonTeamDisplay() {
   const [url, setUrl] = React.useState<string | null>(null);
   const { teamData, loading, error } = usePokePaste(url);
+  const [input, setInput] = React.useState<string>("");
 
   const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
@@ -27,23 +28,30 @@ export default function PokemonTeamDisplay() {
 
   return (
     <div className="container mx-auto px-4">
-      <form className="my-8" onSubmit={handleSubmit}>
-        <Input fullWidth isClearable className="mb-2" name="url" placeholder="Enter PokePaste URL" />
-        <Button color="primary" type="submit">
-          Load Team
-        </Button>
+      <form onSubmit={ handleSubmit } className="my-8">
+        <Textarea
+          name="url"
+          placeholder="Paste your PokePaste URL or Showdown Set here"
+          fullWidth
+          minRows={ 3 }
+          maxRows={ 10 }
+          value={input}
+          onChange={ (e) => setInput(e.target.value) }
+          className="mb-2"
+        />
+        <Button type="submit" color="primary">Load Team</Button>
       </form>
 
-      {loading && <div className="text-center">Loading...</div>}
-      {error && <p className="text-danger">Error: {error.message}</p>}
+      { loading && <div className="text-center">Loading...</div> }
+      { error && <p className="text-danger">Error: { error.message }</p> }
 
-      {teamData && (
+      { teamData && (
         <div className="mb-8">
-          <h1 className="text-2xl font-bold">{teamData.metadata.title}</h1>
-          <p>Author: {teamData.metadata.author}</p>
-          <p>Format: {teamData.metadata.format}</p>
+          <h1 className="text-2xl font-bold">{ teamData.metadata.title || "Custom Team" }</h1>
+          { teamData.metadata.author && <p>Author: { teamData.metadata.author }</p> }
+          { teamData.metadata.format && <p>Format: { teamData.metadata.format }</p> }
         </div>
-      )}
+      ) }
 
       <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4">
         {teamData &&
