+++ frontend/lib/pokemon/use-poke-paste.tsx
@@ -1,35 +1,6 @@
 import { useState, useEffect } from "react";
-import { PokemonSet } from "@pkmn/sets";
-import { StatsTable } from "@pkmn/types";
-
-interface ParsedPokemon extends PokemonSet {
-  imgPokemon: string;
-  imgItem: string;
-  ability: string;
-  level: number;
-  teraType: string;
-  evs: StatsTable;
-  ivs: StatsTable;
-  nature: string;
-  moves: string[];
-}
-
-interface PasteMetadata {
-  title: string;
-  author: string;
-  format: string;
-}
-
-interface ParsedTeam {
-  metadata: PasteMetadata;
-  pokemon: ParsedPokemon[];
-}
-
-interface PokePasteResults {
-  teamData: ParsedTeam | null;
-  loading: boolean;
-  error: Error | null;
-}
+import { parsePokePasteHTML } from "./parse-pokepaste-html";
+import { PokePasteResults, ParsedTeam } from "./common";
 
 export function usePokePaste(url?: string | null): PokePasteResults {
   const [teamData, setTeamData] = useState<ParsedTeam | null>(null);
@@ -50,7 +21,7 @@ export function usePokePaste(url?: string | null): PokePasteResults {
         });
 
         const html = await response.text();
-        const parsedData = parsePokemonTeam(html);
+        const parsedData = parsePokePasteHTML(html);
 
         setTeamData(parsedData);
         setLoading(false);
@@ -77,108 +48,3 @@ export function usePokePaste(url?: string | null): PokePasteResults {
   return { teamData, loading, error };
 }
 
-function parsePokemonTeam(html: string): ParsedTeam {
-  if (typeof window === "undefined") return { metadata: { title: "", author: "", format: "" }, pokemon: [] };
-
-  const parser = new DOMParser();
-  const doc = parser.parseFromString(html, "text/html");
-
-  // Parse metadata
-  const aside = doc.querySelector("aside");
-  const metadata: PasteMetadata = {
-    title: aside?.querySelector("h1")?.textContent?.trim() || "",
-    author:
-      aside
-        ?.querySelector("h2")
-        ?.textContent?.trim()
-        .replace(/^\s*by\s*/, "") || "",
-    format: aside?.querySelector("p")?.textContent?.replace("Format:", "").trim() || "",
-  };
-
-  const articles = doc.querySelectorAll("article");
-
-  const pokemon = Array.from(articles).map((article): ParsedPokemon => {
-    const preContent = (article.querySelector("pre")?.textContent || "").split("\n");
-
-    const lines = preContent.map((line) => line.trim()).filter(Boolean);
-
-    const speciesLine = (lines[0] && lines[0].split(" @ ")) as string[];
-    const species = speciesLine[0] as string;
-    const itemLine = speciesLine[1] ? speciesLine[1].split("\\n") : [];
-    const item = itemLine[0] || "";
-
-    const ability = itemLine.find((line) => line.startsWith("Ability:"))?.split(": ")[1] || "";
-    const level = parseInt(itemLine.find((line) => line.startsWith("Level:"))?.split(": ")[1] || "100", 10);
-    const teraType = itemLine.find((line) => line.startsWith("Tera Type:"))?.split(": ")[1] || "";
-    const nature = itemLine.find((line) => line.includes("Nature"))?.split(" ")[0] || "";
-
-    const moves = itemLine.filter((line) => line.startsWith("-")).map((line) => line.substring(2).trim());
-
-    const evsLine = itemLine.find((line) => line.startsWith("EVs:"))?.split(": ")[1] || "";
-    const ivsLine = itemLine.find((line) => line.startsWith("IVs:"))?.split(": ")[1] || "";
-
-    const imgDiv = article.querySelector("div");
-    const imgPokemonSrc = (imgDiv?.querySelector("img")?.getAttribute("src") || "").replace(/^["']|["']$/g, "");
-    const imgItemSrc = (imgDiv?.querySelectorAll("img")[1]?.getAttribute("src") || "").replace(/^["']|["']$/g, "");
-
-    const imgPokemon = cleanImageUrl(imgPokemonSrc);
-    const imgItem = cleanImageUrl(imgItemSrc);
-
-    return {
-      name: "",
-      species,
-      gender: "",
-      item,
-      ability,
-      level,
-      teraType,
-      evs: parseStats(evsLine, 0),
-      ivs: parseStats(ivsLine),
-      nature,
-      moves,
-      imgPokemon,
-      imgItem,
-    };
-  });
-
-  return { metadata, pokemon };
-}
-
-function cleanImageUrl(url: string): string {
-  // Remove any quotes and extra spaces
-  url = url.replace(/["'\s]/g, "");
-
-  // Ensure the URL starts with a forward slash
-  if (url && !url.startsWith("/")) {
-    url = "/" + url;
-  }
-
-  // Prepend the base URL only if we have a valid path
-  return url ? `https://pokepast.es${url}` : "";
-}
-
-function parseStats(statsLine: string, defaultValue?: number): StatsTable {
-  const statsTable: Partial<StatsTable> = {};
-
-  statsLine.split("/").forEach((stat) => {
-    const [value, name] = stat.trim().split(" ");
-
-    if (name && value) {
-      const statKey = name.toLowerCase() as keyof StatsTable;
-
-      statsTable[statKey] = parseInt(value, 10);
-    }
-  });
-
-  if (defaultValue !== undefined) {
-    const allStats: (keyof StatsTable)[] = ["hp", "atk", "def", "spa", "spd", "spe"];
-
-    allStats.forEach((stat) => {
-      if (statsTable[stat] === undefined) {
-        statsTable[stat] = defaultValue;
-      }
-    });
-  }
-
-  return statsTable as StatsTable;
-}
